% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFPKM.R
\name{calculateFPKMMatrix}
\alias{calculateFPKMMatrix}
\title{calculate FPKM from countTable and coldata}
\usage{
calculateFPKMMatrix(countTable, coldata, geneInfo)
}
\arguments{
\item{countTable}{a raw read count table}

\item{coldata}{an annotation table for samples, must include attributes naming "barcode", "patient_barcode", "survival", "vital_status", "total_count", "cohort"}

\item{geneInfo}{an annotation table for genes, must include attribute nameing "gene_length"}
}
\value{
FPKM matrix
}
\description{
calculate FPKM from countTable and coldata
}
