% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalAnalysis.R
\name{coxSurvivalAnalysis}
\alias{coxSurvivalAnalysis}
\title{Cox survival analysis}
\usage{
coxSurvivalAnalysis(gene, rpkm, coldata)
}
\arguments{
\item{gene}{gene id}

\item{rpkm}{FPKM matrix}

\item{coldata}{an annotation table for samples}
}
\value{
a list containing cox regression p value and coef
}
\description{
Cox survival analysis
}
