% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalAnalysis.R
\name{coxSurvivalAnalysisForGenes}
\alias{coxSurvivalAnalysisForGenes}
\title{Cox survival analysis for a list of genes}
\usage{
coxSurvivalAnalysisForGenes(geneList, rpkm, coldata)
}
\arguments{
\item{geneList}{a list of gene id}

\item{rpkm}{FPKM matrix}

\item{coldata}{coldata an annotation table for samples}
}
\value{
a data frame containing cox regression results (gene id, cox regression p value, coef)
}
\description{
Cox survival analysis for a list of genes
}
