% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAnalysis.R
\name{differentialAnalysis}
\alias{differentialAnalysis}
\title{differential analysis function based on DESeq2}
\usage{
differentialAnalysis(countTable, coldata, fields, geneInfo,
  datasetIdentifier = "")
}
\arguments{
\item{countTable}{a raw read count table}

\item{coldata}{an annotation table for samples, must include attributes naming "barcode", "patient_barcode", "survival", "vital_status", "total_count", "cohort"}

\item{fields}{name list of the attributes for differential expression analysis}

\item{geneInfo}{an annotation table for genes, must include attribute nameing "gene_length"}

\item{datasetIdentifier}{a prefix for the output csv files}
}
\value{
a list of \code{\link{DESeqDataSet}} object and output csv files for differential analysis results and vsd normalized expression table directly to current directory
}
\description{
Do differential analysis with DESeq2 recursively with attributes defined in 'fields' parameter
}
