% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAnalysis.R
\name{differentialAnalysisSingleAttribute}
\alias{differentialAnalysisSingleAttribute}
\title{differential analysis function based on DESeq2
Do differential analysis with DESeq2 on single attribute and two attribute}
\usage{
differentialAnalysisSingleAttribute(countTable, coldata, attribute, geneInfo,
  datasetIdentifier)
}
\arguments{
\item{countTable}{a raw read count table}

\item{coldata}{an annotation table for samples, must include attributes naming "barcode", "patient_barcode", "survival", "vital_status", "total_count", "cohort"}

\item{attribute}{name of the attribute for differential expression analysis}

\item{geneInfo}{an annotation table for genes, must include attribute nameing "gene_length"}

\item{datasetIdentifier}{a prefix for the output csv files}
}
\value{
a \code{\link{DESeqDataSet}} object and output two csv files for differential analysis results and vsd normalized expression table directly to current directory
}
\description{
differential analysis function based on DESeq2
Do differential analysis with DESeq2 on single attribute and two attribute
}
