% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionCorrelation.R
\name{expressionCorrelation}
\alias{expressionCorrelation}
\title{calculate correlation of expression between a gene and a gene list}
\usage{
expressionCorrelation(gene, matrix1, matrix2)
}
\arguments{
\item{gene}{targe gene id}

\item{matrix1}{FPKM matrix containing target gene id}

\item{matrix2}{FPKM matrix for looping and calculate pearson correlation with target gene}
}
\value{
a data.frame containing gene id, pearson coefficient, p value
}
\description{
calculate correlation of expression between a gene and a gene list
}
