% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionCorrelation.R
\name{expressionCorrelationSingleGene}
\alias{expressionCorrelationSingleGene}
\title{calculate correlation of expression between two genes}
\usage{
expressionCorrelationSingleGene(gene1, matrix1, gene2, matrix2)
}
\description{
calculate correlation of expression between two genes
}
