% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFPKM.R
\name{plotFPKMBoxplot}
\alias{plotFPKMBoxplot}
\title{plot RPKM boxplot}
\usage{
plotFPKMBoxplot(FPKMTable, coldata, geneInfo, attribute, geneName = "",
  geneId = "")
}
\arguments{
\item{FPKMTable}{FPKM matrix}

\item{coldata}{an annotation table for samples, must include attributes naming "barcode", "patient_barcode", "survival", "vital_status", "total_count", "cohort"}

\item{geneInfo}{an annotation table for genes, must include attribute nameing "gene_length"}

\item{attribute}{attribute used to group data (x axis)}

\item{geneName}{the name of gene to draw boxplot}

\item{geneId}{the Id of gene to draw boxplot}
}
\value{
boxplot of FPKM by ggplot2
}
\description{
plot RPKM boxplot
}
