% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAnalysis.R
\name{plotHeatMap}
\alias{plotHeatMap}
\title{plot heatmap with vsdData, differentialData}
\usage{
plotHeatMap(vsdData, coldata, orderAttribute, labelAttribute,
  differentialData = NULL, pValueCutoff = 1, averageCutoff = 0,
  varCutoff = 0, clusterRow = TRUE, clusterCol = TRUE, colorBias = 1,
  showNames = FALSE, rowFontSize = 2, colFontSize = 3)
}
\value{
heatmap by pheatmap
}
\description{
plot heatmap with vsdData, differentialData
}
