% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalAnalysis.R
\name{plotSurvivalAnalysis}
\alias{plotSurvivalAnalysis}
\title{plot survival curves}
\usage{
plotSurvivalAnalysis(gene, rpkm, coldata, topn = 10, rpkmCutoff = 0)
}
\arguments{
\item{gene}{id}

\item{coldata}{an annotation table for samples}

\item{topn}{group samples into two groups: high expression and low expression. Classify the top n samples (by FPKM) into high expression group.}

\item{rpkmCutoff}{classify samples into two groups by a rpkmCutoff}
}
\value{
a list containing survival plot, p value and table for plotting
}
\description{
plot survival curves
}
