% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalAnalysis.R
\name{prepareTableForSurvivalAnalysis}
\alias{prepareTableForSurvivalAnalysis}
\title{Prepare table for survival analysis}
\usage{
prepareTableForSurvivalAnalysis(gene, rpkm, coldata)
}
\description{
Prepare table for survival analysis
}
